Move/refactory sync_* methods
